Generalize finding tool sheds from runnable
Should now work for workflow_edit, test, run etc if targeting a local or
docker instance.
