+++ planemo/galaxy/workflows.py
@@ -61,6 +61,14 @@ def guess_tool_shed_url(tool_shed_fqdn: str) -> Optional[str]:
     return None
 
 
+def get_toolshed_url_for_tool_id(tool_id: str) -> Optional[str]:
+    components = tool_id.split("/repos")
+    if len(components) > 1:
+        tool_shed_fqdn = components[0]
+        return guess_tool_shed_url(tool_shed_fqdn=tool_shed_fqdn)
+    return None
+
+
 def load_shed_repos(runnable):
     if runnable.type.name != "galaxy_workflow":
         return []
@@ -164,20 +172,25 @@ def _raw_dict(path, importer=None):
     return workflow
 
 
+def get_tool_ids_for_workflow(wf_dict: Dict[str, Any], tool_ids: Optional[List[str]] = None) -> List[str]:
+    tool_ids = [] if tool_ids is None else tool_ids
+    steps = wf_dict["steps"].values() if isinstance(wf_dict["steps"], dict) else wf_dict["steps"]
+    for step in steps:
+        if step.get("type", "tool") == "tool" and not step.get("run", {}).get("class") == "GalaxyWorkflow":
+            tool_id = step["tool_id"]
+            tool_ids.append(tool_id)
+        elif step.get("type") == "subworkflow":  # GA SWF
+            get_tool_ids_for_workflow(step["subworkflow"], tool_ids=tool_ids)
+        elif step.get("run", {}).get("class") == "GalaxyWorkflow":  # gxformat2 SWF
+            get_tool_ids_for_workflow(step["run"], tool_ids=tool_ids)
+        else:
+            continue
+    return list(dict.fromkeys(tool_ids))
+
+
 def find_tool_ids(path):
-    tool_ids = set()
     workflow = _raw_dict(path)
-
-    def register_tool_ids(tool_ids, workflow):
-        for step in workflow["steps"].values():
-            if step.get("subworkflow"):
-                register_tool_ids(tool_ids, step["subworkflow"])
-            elif step.get("tool_id"):
-                tool_ids.add(step["tool_id"])
-
-    register_tool_ids(tool_ids, workflow)
-
-    return list(tool_ids)
+    return get_tool_ids_for_workflow(workflow)
 
 
 WorkflowOutput = namedtuple("WorkflowOutput", ["order_index", "output_name", "label", "optional"])
